requirements.txt
albumentations==0.5.2
blis==0.7.9
catalogue==2.0.8
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
cycler==0.11.0
cymem==2.0.7
efficientnet-pytorch==0.6.3
fastai==2.7.10
fastcore==1.5.27
fastdownload==0.0.7
fastprogress==1.0.3
fonttools==4.38.0
idna==3.4
imageio==2.24.0
imgaug==0.4.0
importlib-metadata==6.0.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.3
munch==2.5.0
murmurhash==1.0.9
networkx==2.6.3
numpy==1.21.6
opencv-python==4.2.0.34
opencv-python-headless==4.7.0.68
packaging==23.0
pandas==0.25.3
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
pretrainedmodels==0.7.4
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-ignite==0.4.10
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
segmentation-models-pytorch==0.1.3
shapely==2.0.0
six==1.16.0
smart-open==6.3.0
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
srsly==2.4.5
thinc==8.1.7
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.3.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.1.1
urllib3==1.26.14
wasabi==0.10.1
yacs==0.1.8
zipp==3.11.0
